% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{.resolve_backend}
\alias{.resolve_backend}
\title{Resolve backend based on installed packages and GPU availability}
\usage{
.resolve_backend(backend = c("auto", "gpu", "cpu-fast", "cpu-pure"))
}
\description{
Resolve backend based on installed packages and GPU availability
}
\keyword{internal}
