% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addup.R
\name{hello}
\alias{hello}
\title{FUNCTION_TITLE}
\usage{
hello(n)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
