% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecAct-package.R
\docType{package}
\name{SecAct-package}
\alias{SecAct}
\alias{SecAct-package}
\title{SecAct: Secreted Signaling Activity Inference}
\description{
Inferring secreted protein activities at bulk, single-cell, and spatial levels
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/data2intelligence/SecAct}
  \item Report bugs at \url{https://github.com/data2intelligence/SecAct/issues}
}

}
\author{
\strong{Maintainer}: Beibei Ru \email{beibei.ru@nih.gov}

}
\keyword{internal}
