% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.activity.inference.scRNAseq.2}
\alias{SecAct.activity.inference.scRNAseq.2}
\title{Cell type activity inference between two conditions from single cell data}
\usage{
SecAct.activity.inference.scRNAseq.2(
  inputProfile,
  cellType_meta,
  condition_meta,
  conditionCase,
  conditionControl,
  sigMatrix = "SecAct",
  lambda = 5e+05,
  nrand = 1000,
  sigFilter = FALSE
)
}
\arguments{
\item{inputProfile}{A Seurat object.}

\item{cellType_meta}{Column name in meta data that includes cell-type annotations.}

\item{condition_meta}{Column name in meta data that includes condition information.}

\item{conditionCase}{Case condition.}

\item{conditionControl}{Control condition.}

\item{sigMatrix}{Secreted protein signature matrix.}

\item{lambda}{Penalty factor in the ridge regression.}

\item{nrand}{Number of randomization in the permutation test, with a default value 1000.}

\item{sigFilter}{A logical indicating whether filter the secreted protein signatures with the genes from inputProfile.}
}
\value{
A Seurat object.
}
\description{
Calculate the changes in secreted protein signaling activity for each cell type between two conditions from single cell RNA-Sequencing data.
}
