% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.activity.inference.scRNAseq}
\alias{SecAct.activity.inference.scRNAseq}
\title{Cell state activity inference from single cell data}
\usage{
SecAct.activity.inference.scRNAseq(
  inputProfile,
  cellType_meta,
  sigMatrix = "SecAct",
  is.group.sig = TRUE,
  lambda = 5e+05,
  nrand = 1000,
  sigFilter = FALSE
)
}
\arguments{
\item{inputProfile}{A Seurat object.}

\item{cellType_meta}{Column name in meta data that includes cell-type annotations.}

\item{sigMatrix}{Secreted protein signature matrix.}

\item{is.group.sig}{A logical indicating whether to group similar signatures.}

\item{lambda}{Penalty factor in the ridge regression.}

\item{nrand}{Number of randomization in the permutation test, with a default value 1000.}

\item{sigFilter}{A logical indicating whether filter the secreted protein signatures with the genes from inputProfile.}
}
\value{
A Seurat object.
}
\description{
Calculate secreted protein signaling activity of cell states from single cell RNA-Sequencing data.
}
