% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.activity.inference.ST}
\alias{SecAct.activity.inference.ST}
\title{Spot activity inference from spatial data}
\usage{
SecAct.activity.inference.ST(
  inputProfile,
  inputProfile_control = NULL,
  scale.factor = 1e+05,
  sigMatrix = "SecAct",
  is.filter.sig = FALSE,
  is.group.sig = TRUE,
  is.group.cor = 0.9,
  lambda = 5e+05,
  nrand = 1000
)
}
\arguments{
\item{inputProfile}{A SpaCET object.}

\item{inputProfile_control}{A SpaCET object.}

\item{scale.factor}{Sets the scale factor for spot-level normalization.}

\item{sigMatrix}{Secreted protein signature matrix. Could be "SecAct", "CytoSig", "SecAct-Breast", "SecAct-Colorectal", "SecAct-Glioblastoma", "SecAct-Kidney", "SecAct-Liver", "SecAct-Lung-Adeno", "SecAct-Ovarian", "SecAct-Pancreatic", "SecAct-Prostate". SecAct signatures were derived from all cancer ST samples; SecAct-XXXX signatures were derived from XXXX cancer ST samples.}

\item{is.filter.sig}{A logical flag indicating whether to filter the secreted protein signatures based on the genes from inputProfile (Default: FALSE). Because some sequencing platforms (e.g., CosMx) cover only a subset of secreted proteins, setting this option to TRUE restricts the activity inference on those proteins.}

\item{is.group.sig}{A logical flag indicating whether to group similar signatures (Default: TRUE). Many secreted proteins, such as cytokines with similar cell surface receptors and downstream pathways, have cellular effects that appear redundant within a cellular context. When enabled, this option clusters secreted proteins based on Pearson correlations among their composite signatures. The output still reports activity estimates for all secreted proteins prior to clustering. Secreted proteins assigned to the same non-redundant cluster share the same inferred activity.}

\item{is.group.cor}{A numeric value specifying the correlation cutoff used to define similar signatures (Default: 0.90). When r > 0.90, 1,170 secreted protein signatures are grouped into 657 non-redundant signature groups.}

\item{lambda}{Penalty factor in the ridge regression. If NULL, lambda will be assigned as 5e+05 or 10000 when sigMatrix = "SecAct" or "CytoSig", respectively.}

\item{nrand}{Number of randomization in the permutation test, with a default value 1000.}
}
\value{
A SpaCET object.
}
\description{
Calculate secreted protein signaling activity of spots from spatial transcriptomocs data.
}
