% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SecAct.bar.plot}
\alias{SecAct.bar.plot}
\title{Draw a bar plot}
\usage{
SecAct.bar.plot(fg.vec, title = NULL, colors = c("#91bfdb", "#fc8d59"))
}
\arguments{
\item{fg.vec}{A vector of values.}

\item{title}{The title for plot.}

\item{colors}{Colors.}
}
\value{
A ggplot2 object.
}
\description{
Draw a bar plot of secreted proteins.
}
