% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SecAct.CCC.circle}
\alias{SecAct.CCC.circle}
\title{Cell-cell communication circle plot}
\usage{
SecAct.CCC.circle(data, colors_cellType)
}
\arguments{
\item{data}{A SpaCET object or a Seurat object.}

\item{colors_cellType}{Colors for cell types.}
}
\value{
A ggplot2 object.
}
\description{
Draw a circle plot of cell-cell communicationss mediated by secreted proteins.
}
