% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SecAct.CCC.dot}
\alias{SecAct.CCC.dot}
\title{Cell-cell communication dot plot}
\usage{
SecAct.CCC.dot(data, sender = NULL, secretedProtein = NULL, receiver = NULL)
}
\arguments{
\item{data}{A SpaCET object or a Seurat object.}

\item{sender}{Sender cell types.}

\item{secretedProtein}{Secreted proteins.}

\item{receiver}{Receiver cell types.}
}
\value{
A ggplot2 object.
}
\description{
Draw a dot plot of cell-cell communication mediated by secreted proteins.
}
