% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SecAct.CCC.heatmap}
\alias{SecAct.CCC.heatmap}
\title{Cell-cell communication heatmap}
\usage{
SecAct.CCC.heatmap(
  data,
  row.sorted = FALSE,
  column.sorted = FALSE,
  colors_cellType
)
}
\arguments{
\item{data}{A SpaCET object or a Seurat object.}

\item{row.sorted}{Whether to sort rows.}

\item{column.sorted}{Whether to sort columns.}

\item{colors_cellType}{Colors for cell types.}
}
\value{
A ggplot2 object.
}
\description{
Draw a heatmap of cell-cell communicationss mediated by secreted proteins.
}
