% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SecAct.CCC.sankey}
\alias{SecAct.CCC.sankey}
\title{Cell-cell communication sankey plot}
\usage{
SecAct.CCC.sankey(
  data,
  colors_cellType,
  sender = NULL,
  secretedProtein = NULL,
  receiver = NULL
)
}
\arguments{
\item{data}{A SpaCET object or a Seurat object.}

\item{colors_cellType}{Colors for cell types.}
}
\value{
A ggplot2 object.
}
\description{
Draw a sankey plot of cell-cell communication mediated by secreted proteins.
}
