% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.CCC.scRNAseq}
\alias{SecAct.CCC.scRNAseq}
\title{Cell-cell communication from single cell data}
\usage{
SecAct.CCC.scRNAseq(
  Seurat_obj,
  cellType_meta,
  condition_meta = NULL,
  conditionCase = NULL,
  conditionControl = NULL,
  scale.factor = 1e+05,
  act_diff_cutoff = 2,
  exp_logFC_cutoff = 0.2,
  exp_mean_all_cutoff = 2,
  exp_fraction_case_cutoff = 0.1,
  padj_cutoff = 0.01,
  sigMatrix = "SecAct",
  is.group.sig = TRUE,
  is.group.cor = 0.9,
  lambda = 5e+05,
  nrand = 1000
)
}
\arguments{
\item{Seurat_obj}{A Seurat object.}

\item{cellType_meta}{Column name in meta data that includes cell-type annotations.}

\item{condition_meta}{Column name in meta data that includes condition information.}

\item{conditionCase}{Case condition.}

\item{conditionControl}{Control condition.}

\item{scale.factor}{Sets the scale factor for cell-level normalization in step2.}

\item{act_diff_cutoff}{Cut off for activity change (i.e., z score) in step 1.}

\item{exp_logFC_cutoff}{Cut off for log fold change in step 2.}

\item{exp_mean_all_cutoff}{Cut off for mean expression across all cells required in step 2.}

\item{exp_fraction_case_cutoff}{Cut off for the fraction of cells expressing secreted protein-coding genes in step 2.}

\item{padj_cutoff}{Adjusted p value cut off.}

\item{sigMatrix}{Secreted protein signature matrix.}

\item{is.group.sig}{A logical flag indicating whether to group similar signatures.}

\item{is.group.cor}{Correlation cutoff of similar signatures.}

\item{lambda}{Penalty factor in the ridge regression.}

\item{nrand}{Number of randomization in the permutation test, with a default value 1000.}
}
\value{
A Seurat object.
}
\description{
Calculate condition-specific cell-cell communication mediated by secreted proteins from scRNA-Seq data.
}
