% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.CCC.scST}
\alias{SecAct.CCC.scST}
\title{Cell-cell communication from spatial data}
\usage{
SecAct.CCC.scST(data, cellType_meta, radius = 0.02, padj_cutoff = 0.01)
}
\arguments{
\item{data}{A SpaCET object.}

\item{cellType_meta}{Column name in meta data that includes cell-type annotations.}

\item{padj_cutoff}{Adjusted p value cut off.}
}
\value{
A Seurat object.
}
\description{
Calculate cell-cell communication mediated by secreted proteins from spatial transcriptomics data.
}
