% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.coxph}
\alias{SecAct.coxph}
\title{Survival regression}
\usage{
SecAct.coxph(mat, surv)
}
\arguments{
\item{mat}{Activity matrix.}

\item{surv}{Survival matrix.}
}
\value{
A matrix.
}
\description{
Calculate the risk score of each secreted protein.
}
