% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.inference.gsl}
\alias{SecAct.inference.gsl}
\title{Secreted protein activity inference}
\usage{
SecAct.inference.gsl(Y, SigMat = "SecAct", lambda = 5e+05, nrand = 1000)
}
\arguments{
\item{Y}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{SigMat}{Secreted protein signature matrix.}

\item{lambda}{Penalty factor in the ridge regression.}

\item{nrand}{Number of randomizations in the permutation test, with a default value 1000.}
}
\value{
A list with four items, each is a matrix.
beta: regression coefficients
se: standard errors of coefficients
zscore: beta/se
pvalue: statistical significance
}
\description{
Infer the activity of over 1000 secreted proteins from tumor gene expression profiles.
}
