% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.inference}
\alias{SecAct.inference}
\title{Secreted protein activity inference}
\usage{
SecAct.inference(expr, SigMat = "SecAct", lambda = 5e+05, nrand = 1000)
}
\arguments{
\item{expr}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{lambda}{Penalty factor in the ridge regression.}

\item{nrand}{Number of randomizations in the permutation test, with a default value 1000.}

\item{sigMatrix}{Secreted protein signature matrix.}
}
\value{
A list with four items, each is a matrix.
beta: regression coefficients
se: standard errors of coefficients
zscore: beta/se
pvalue: statistical significance
}
\description{
Infer the activity of over 1000 secreted proteins from tumor gene expression profiles.
}
\examples{

exprPath <- file.path(system.file(package = "SecAct"), "extdata/IFNG_GSE100093.diff")
expr <- read.table(expr, sep="\t", check.names=F)
res <- SecAct.inference(expr, lambda=10000, nrand=1000)
head(res$zscore)

}
