% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SecAct.lollipop.plot}
\alias{SecAct.lollipop.plot}
\title{Draw a lollipop plot}
\usage{
SecAct.lollipop.plot(fg.vec, title = NULL)
}
\arguments{
\item{fg.vec}{A vector of values.}

\item{title}{The title for plot.}
}
\value{
A ggplot2 object.
}
\description{
Draw a lollipop plot of secreted proteins.
}
