% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.signaling.velocity.spotST}
\alias{SecAct.signaling.velocity.spotST}
\title{Secreted protein signaling velocity}
\usage{
SecAct.signaling.velocity.spotST(
  SpaCET_obj,
  scale.factor = 1e+05,
  gene,
  signalMode = "receiving",
  radius = 200,
  contourMap = FALSE,
  contourBins = 11,
  animated = FALSE
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{scale.factor}{Sets the scale factor for spot-level normalization.}

\item{gene}{Gene symbol coding a secreted protein.}

\item{signalMode}{Mode of signaling velocity, i.e., "receiving" or "sending".}

\item{radius}{Radius cut off.}

\item{contourMap}{A logical flag indicating whether to transform as contour map.}

\item{contourBins}{Number of bins used to generate the contour map.}

\item{animated}{A logical flag indicating whether to generate animated figure.}
}
\value{
A ggplot2 object.
}
\description{
Calculate the signaling velocity of secreted proteins based on their activities.
}
\details{
The velocity direction starts from the source cell producing a secreted protein and moves to sink cells receiving the secreted protein signal. The velocity magnitude represents the product between the secreted protein-coding gene expression at source cells and signaling activities at sink cells.
}
\examples{
SecAct.signaling.velocity.spotST(SpaCET_obj, gene="TGFB1", signalMode="receiving")
SecAct.signaling.velocity.spotST(SpaCET_obj, gene="TGFB1", signalMode="sending")

}
