% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SecAct.signalling.direction}
\alias{SecAct.signalling.direction}
\title{FUNCTION_TITLE}
\usage{
SecAct.signalling.direction(SpaCET_obj, gene = "TGFB1")
}
\arguments{
\item{SpaCET_obj}{PARAM_DESCRIPTION}

\item{gene}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
