% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SecAct.spatial.velocity}
\alias{SecAct.spatial.velocity}
\title{Secreted protein signaling velocity}
\usage{
SecAct.spatial.velocity(SpaCET_obj, gene, signalMode = "both")
}
\arguments{
\item{SpaCET_obj}{An SpaCET object.}

\item{gene}{Gene symbol coding a secreted protein.}

\item{signalMode}{Mode of signaling velocity, i.e., "receiving", "sending", and "both".}
}
\value{
A ggplot2 object.
}
\description{
Calculate the signaling velocity of secreted proteins based on their activities.
}
\details{
The velocity direction starts from the source cell producing a secreted protein and moves to sink cells receiving the secreted protein signal. The velocity magnitude represents the product between the secreted protein-coding gene expression at source cells and signaling activities at sink cells.
}
\examples{
SecAct.spatial.velocity(SpaCET_obj, gene="TGFB1", signalMode="receiving")
SecAct.spatial.velocity(SpaCET_obj, gene="TGFB1", signalMode="sending")
SecAct.spatial.velocity(SpaCET_obj, gene="TGFB1", signalMode="both")

}
