% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{SecAct.survival.plot}
\alias{SecAct.survival.plot}
\title{Draw a survival plot}
\usage{
SecAct.survival.plot(mat, surv, gene, x.title = "Time")
}
\arguments{
\item{mat}{Activity matrix.}

\item{surv}{Survival matrix.}

\item{gene}{Gene symbol.}

\item{x.title}{Title for x axis.}
}
\value{
A ggplot2 object.
}
\description{
Draw a survival plot of secreted proteins.
}
