% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.activity.inference}
\alias{SecAct.activity.inference}
\title{Secreted protein activity inference}
\usage{
SecAct.activity.inference(
  inputProfile,
  inputProfile_control = NULL,
  is.differential = FALSE,
  is.paired = FALSE,
  is.singleSampleLevel = FALSE,
  sigMatrix = "SecAct",
  is.filter.sig = FALSE,
  is.group.sig = TRUE,
  is.group.cor = 0.9,
  lambda = 5e+05,
  nrand = 1000
)
}
\arguments{
\item{inputProfile}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{inputProfile_control}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{is.differential}{A logical flag indicating whether inputProfile has been differential profiles against the control (Default: FALSE).}

\item{is.paired}{A logical flag indicating whether you want a paired operation of differential profiles between inputProfile and inputProfile_control if samples in inputProfile and inputProfile_control are paired (Default: FALSE).}

\item{is.singleSampleLevel}{A logical flag indicating whether to calculate activity change for each single sample between inputProfile and inputProfile_control (Default: FALSE). If FALSE, calculate the overall activity change between two phenotypes .}

\item{sigMatrix}{Secreted protein signature matrix. Could be "SecAct", "CytoSig", "SecAct-Breast", "SecAct-Colorectal", "SecAct-Glioblastoma", "SecAct-Kidney", "SecAct-Liver", "SecAct-Lung-Adeno", "SecAct-Ovarian", "SecAct-Pancreatic", "SecAct-Prostate". SecAct signatures were derived from all cancer ST samples; SecAct-XXXX signatures were derived from XXXX cancer ST samples.}

\item{is.filter.sig}{A logical flag indicating whether to filter the secreted protein signatures based on the genes from inputProfile (Default: FALSE). Because some sequencing platforms (e.g., CosMx) cover only a subset of secreted proteins, setting this option to TRUE restricts the activity inference on those proteins.}

\item{is.group.sig}{A logical flag indicating whether to group similar signatures (Default: TRUE). Many secreted proteins, such as cytokines with similar cell surface receptors and downstream pathways, have cellular effects that appear redundant within a cellular context. When enabled, this option clusters secreted proteins based on Pearson correlations among their composite signatures. The output still reports activity estimates for all secreted proteins prior to clustering. Secreted proteins assigned to the same non-redundant cluster share the same inferred activity.}

\item{is.group.cor}{A numeric value specifying the correlation cutoff used to define similar signatures (Default: 0.90). When r > 0.90, 1,170 secreted protein signatures are grouped into 657 non-redundant signature groups.}

\item{lambda}{Penalty factor in the ridge regression. If NULL, lambda will be assigned as 5e+05 or 10000 when sigMatrix = "SecAct" or "CytoSig", respectively.}

\item{nrand}{Number of randomization in the permutation test, with a default value of 1000.}
}
\value{
A list with four items, each is a matrix.
beta: regression coefficients
se: standard errors of coefficients
zscore: beta/se
pvalue: statistical significance
}
\description{
Infer the signaling activity of over 1000 secreted proteins from gene expression profiles.
}
