% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.activity.inference}
\alias{SecAct.activity.inference}
\title{Secreted protein activity inference}
\usage{
SecAct.activity.inference(
  inputProfile,
  inputProfile_control = NULL,
  is.differential = FALSE,
  is.paired = FALSE,
  is.singleSampleLevel = FALSE,
  sigMatrix = "SecAct",
  is.group.sig = TRUE,
  is.group.cor = 0.9,
  lambda = 5e+05,
  nrand = 1000,
  sigFilter = FALSE
)
}
\arguments{
\item{inputProfile}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{inputProfile_control}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{is.differential}{A logical indicating whether inputProfile has been differential profiles against to control.}

\item{is.paired}{A logical indicating whether you want a paired operation of differential profiles between inputProfile and inputProfile_control if samples in inputProfile and inputProfile_control are paired.}

\item{is.singleSampleLevel}{A logical indicating whether to calculate activity change for each single sample between inputProfile and inputProfile_control. If FALSE, calculate the overall activity change between two phenotypes.}

\item{sigMatrix}{Secreted protein signature matrix.}

\item{is.group.sig}{A logical indicating whether group similar signatures.}

\item{is.group.cor}{Correlation cutoff of similar signatures.}

\item{lambda}{Penalty factor in the ridge regression.}

\item{nrand}{Number of randomization in the permutation test, with a default value 1000.}

\item{sigFilter}{A logical indicating whether filter the secreted protein signatures with the genes from inputProfile.}
}
\value{
A list with four items, each is a matrix.
beta: regression coefficients
se: standard errors of coefficients
zscore: beta/se
pvalue: statistical significance
}
\description{
Infer the signaling activity of 1248 secreted proteins from gene expression profiles.
}
