% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{SecAct.activity.inference.ST}
\alias{SecAct.activity.inference.ST}
\title{Spot activity inference from spatial data}
\usage{
SecAct.activity.inference.ST(
  inputProfile,
  inputProfile_control = NULL,
  scale.factor = 1e+05,
  sigMatrix = "SecAct",
  is.group.sig = TRUE,
  lambda = 5e+05,
  nrand = 1000,
  sigFilter = FALSE
)
}
\arguments{
\item{inputProfile}{A SpaCET object.}

\item{inputProfile_control}{A SpaCET object.}

\item{scale.factor}{Sets the scale factor for spot-level normalization.}

\item{sigMatrix}{Secreted protein signature matrix.}

\item{is.group.sig}{A logical indicating whether to group similar signatures.}

\item{lambda}{Penalty factor in the ridge regression.}

\item{nrand}{Number of randomization in the permutation test, with a default value 1000.}

\item{sigFilter}{A logical indicating whether filter the secreted protein signatures with the genes from inputProfile.}
}
\value{
A SpaCET object.
}
\description{
Calculate secreted protein signaling activity of spots from spatial transcriptomocs data.
}
