% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.CCC.scRNAseq}
\alias{SecAct.CCC.scRNAseq}
\title{Cell-cell communication from single cell data}
\usage{
SecAct.CCC.scRNAseq(
  Seurat_obj,
  cellType_meta,
  condition_meta,
  conditionCase,
  conditionControl,
  scale.factor = 1e+05,
  act_diff_cutoff = 2,
  exp_logFC_cutoff = 0.2,
  exp_fraction_case_cutoff = 0.1,
  padj_cutoff = 0.01
)
}
\arguments{
\item{cellType_meta}{Column name in meta data that includes cell-type annotations.}

\item{condition_meta}{Column name in meta data that includes condition information.}

\item{conditionCase}{Case condition.}

\item{conditionControl}{Control condition.}

\item{act_diff_cutoff}{Cut off for activity change (i.e., z score) in step 1.}

\item{exp_logFC_cutoff}{Cut off for log fold change in step 2.}

\item{exp_fraction_case_cutoff}{Cut off for the fraction of cells expressing secreted protein-coding genes in step 2.}

\item{padj_cutoff}{Adjusted p value cut off.}

\item{data}{A Seurat object.}
}
\value{
A Seurat object.
}
\description{
Calculate condition-specific cell-cell communication mediated by secreted proteins from scRNA-Seq data.
}
