% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.CCC.scST}
\alias{SecAct.CCC.scST}
\title{Cell-cell communication from spatial data}
\usage{
SecAct.CCC.scST(
  SpaCET_obj,
  cellType_meta,
  scale.factor = 1000,
  radius = 20,
  ratio_cutoff = 0.2,
  padj_cutoff = 0.01,
  coreNo = 6
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{cellType_meta}{Column name in meta data that includes cell-type annotations.}

\item{scale.factor}{Sets the scale factor for spot-level normalization.}

\item{radius}{Radius cut off (unit: um).}

\item{ratio_cutoff}{Ratio cut off.}

\item{padj_cutoff}{Adjusted p value cut off.}

\item{coreNo}{Core number in parallel computation.}
}
\value{
A Seurat object.
}
\description{
Calculate cell-cell communication mediated by secreted proteins from spatial transcriptomics data.
}
