% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addup.R
\name{SecAct.inference}
\alias{SecAct.inference}
\title{FUNCTION_TITLE}
\usage{
SecAct.inference(Y, lambda, nrand)
}
\arguments{
\item{Y}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION}

\item{nrand}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
Yfile<- file.path(system.file(package = "SecAct"), "extdata/IFNG_GSE100093.diff")
Y <- read.table(Yfile,sep="\t",check.names=F)
res <- SecAct.inference(Y, lambda=10000, nrand=1000)
names(res)
head(res$zscore)
}
