% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.signaling.pattern.gene}
\alias{SecAct.signaling.pattern.gene}
\title{Pattern-associated secreted proteins}
\usage{
SecAct.signaling.pattern.gene(SpaCET_obj, n)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{n}{Pattern order.}
}
\value{
A matrix.
}
\description{
Enumerate secreted proteins associated with each signaling pattern.
}
\examples{
SpaCET_obj <- SecAct.pattern.gene(SpaCET_obj, n=3)

}
