% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.signaling.pattern}
\alias{SecAct.signaling.pattern}
\title{Secreted protein signaling pattern}
\usage{
SecAct.signaling.pattern(SpaCET_obj, k = 3)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{k}{Number of patterns for NMF.}
}
\value{
A ggplot2 object.
}
\description{
Calculate the signaling pattern of secreted proteins based on their activities.
}
\examples{
SpaCET_obj <- SecAct.signaling.pattern(SpaCET_obj, k=3)

}
