% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.signaling.velocity.scST}
\alias{SecAct.signaling.velocity.scST}
\title{Secreted protein signaling velocity}
\usage{
SecAct.signaling.velocity.scST(
  SpaCET_obj,
  sender,
  secretedProtein,
  receiver,
  cellType_meta,
  scale.factor = 1e+05,
  CustomizedAreaCoordinates = NULL,
  radius = 20,
  show.coordinates = TRUE,
  colors = NULL,
  pointSize = 1,
  pointAlpha = 1,
  legend.position = "right",
  legend.size = 1,
  arrow.color = "#ff0099",
  arrow.width = 1,
  arrow.size = 0.3
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{sender}{Sender cell types.}

\item{secretedProtein}{Secreted proteins.}

\item{receiver}{Receiver cell types.}

\item{scale.factor}{Sets the scale factor for spot-level normalization.}

\item{CustomizedAreaCoordinates}{A vector of four numbers for coordinates of the Customized Area, i.e., x_left, x_right, y_bottom, y_top.}

\item{radius}{Radius cut off (unit: um).}

\item{show.coordinates}{Whether to show coordinates.}

\item{colors}{A vector of colors for spots.}

\item{pointSize}{Size of spots.}

\item{pointAlpha}{Alpha transparency scales of spots. Must lie between 0 and 1.}

\item{legend.position}{The position of the legend. Set it as "none" if you want to remove the legend.}

\item{legend.size}{The size of the legend.}

\item{arrow.color}{The color of the arrow.}

\item{arrow.width}{The width of the arrow.}

\item{arrow.size}{The size of the arrow.}
}
\value{
A ggplot2 object.
}
\description{
Calculate the signaling velocity of secreted proteins based on their activities.
}
\details{
The velocity direction starts from the source cell producing a secreted protein and moves to sink cells receiving the secreted protein signal. The velocity magnitude represents the product between the secreted protein-coding gene expression at source cells and signaling activities at sink cells.
}
\examples{
SecAct.signaling.velocity.scST(SpaCET_obj, sender="Fibroblast", secretedProtein="THBS2", receiver="Tumor_boundary", cellType_meta="cellType")
SecAct.signaling.velocity.scST(SpaCET_obj, sender="Fibroblast", secretedProtein="THBS2", receiver="Tumor_boundary", cellType_meta="cellType", CustomizedAreaCoordinates=c(8290,8366,1100,1400))

}
