% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{SecAct.signaling.velocity.scST}
\alias{SecAct.signaling.velocity.scST}
\title{Secreted protein signaling velocity}
\usage{
SecAct.signaling.velocity.scST(
  SpaCET_obj,
  sender,
  secretedProtein,
  receiver,
  cellType_meta,
  scale.factor = 1e+05,
  radius = 0.02
)
}
\arguments{
\item{SpaCET_obj}{A SpaCET object.}

\item{sender}{Sender cell types.}

\item{secretedProtein}{Secreted proteins.}

\item{receiver}{Receiver cell types.}

\item{scale.factor}{Sets the scale factor for spot-level normalization.}
}
\value{
A ggplot2 object.
}
\description{
Calculate the signaling velocity of secreted proteins based on their activities.
}
\details{
The velocity direction starts from the source cell producing a secreted protein and moves to sink cells receiving the secreted protein signal. The velocity magnitude represents the product between the secreted protein-coding gene expression at source cells and signaling activities at sink cells.
}
\examples{
SecAct.signaling.velocity.scST(SpaCET_obj, sender="Fibroblast", secretedProtein="THBS2", receiver="Tumor_boundary", cellType_meta="cellType")

}
